requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
jinja2==3.1.2
pydantic==2.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
